src/aggregator/aggregation/counter_pool_test.go
// Copyright (c) 2016 Uber Technologies, Inc.
//
// Permission is hereby granted, free of charge, to any person obtaining a copy
// of this software and associated documentation files (the "Software"), to deal
// in the Software without restriction, including without limitation the rights
// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
// copies of the Software, and to permit persons to whom the Software is
// furnished to do so, subject to the following conditions:
//
// The above copyright notice and this permission notice shall be included in
// all copies or substantial portions of the Software.
//
// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
// THE SOFTWARE.

package aggregation

import (
	"testing"

	"github.com/m3db/m3x/pool"

	"github.com/stretchr/testify/require"
)

func TestCounterPool(t *testing.T) {
	p := NewCounterPool(pool.NewObjectPoolOptions().SetSize(1))
	p.Init()

	// Retrieve a counter from the pool
	counter := p.Get()
	counter.Reset()
	counter.Add(1)
	require.Equal(t, int64(1), counter.Sum())

	// Put the counter back to pool
	p.Put(counter)

	// Retrieve the counter and assert it's the same counter
	counter = p.Get()
	require.Equal(t, int64(1), counter.Sum())

	// Reset the counter and assert it's been reset
	counter.Reset()
	require.Equal(t, int64(0), counter.Sum())
}
